backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
